Support for Unicode paths on Windows
Replaces all file functions with wrappers that map internal UTF-8 paths from
and to wide (UTF-16) paths when interacting with Windows.

UTF-8 to UTF-16 conversion for paths is handled by utf8_to_wbuf(), which
rotates 4 static buffers. We need at least 2 for symlink(), rename() etc.,
so we use 4 to be on the safe side. Other conversions are handled by
utf8_to_wchar_auto(), which returns an allocated buffer that must be freed.

UTF-16 to UTF-8 conversion is handled by wchar_to_utf8() which works like
wcstombs and writes directly to a destination buffer.

Also replaces get_file_attr (which uses GetFileAttributesA) in lstat() with
the slightly faster FindFirstFileW, which makes the symlink handling much
cleaner and avoids an additional FindFirstFile.

Signed-off-by: Nico Rieck <nico.rieck@gmail.com>