Support for Unicode command line arguments on Windows
This means full support for specifying Unicode paths and other arguments when
invoking git on Windows. Without this it's impossible to work with Unicode
paths since all pathspecs are already garbled when git gets them.

Because mingw does not support Unicode command line arguments via
wmain(int, wchar_t**) we have to use GetCommandLineW() and convert it to UTF-8
ourselves. This is done by convert_command_line() which also makes sure that
argv[0] is a full path.

Signed-off-by: Nico Rieck <nico.rieck@gmail.com>