Support for Unicode environment variables on Windows
Replaces environment variable interaction with wrappers that use the wide
function variants to get and set UTF-16 values.

getenv() usually returns a pointer to the environment table and thus there is
no need to free the returned strings. Because we cannot force callers to free
our converted buffer we have to manage this ourselves. Therefore we store
converted values in a hashtable (env_map) and release them again when the
variable is cleared or changed.

Signed-off-by: Nico Rieck <nico.rieck@gmail.com>