Tentative workaround for issue 543
Of course, this is only the first step to squeeze the reporter's pimple.
The next step would be to find out why the whole dance around getcwd()
vs getenv("PWD") is needed in the first place, and why the inode is
compared.

This patch only addresses the issue that the inode field is always set
to 0 on msysGit and cannot be used for reliable comparison.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>